# swagger.json
{
    "openapi": "3.0.1",
    "info": {
    "title": "Music App API",
    "description": "API documentation for Music App",
    "version": "1.0"
    },
    "servers": [
    {
    "url": "http://localhost:9000",
    "description": "Generated server url"
    }
    ],
...
    "paths": {
...
    "/artists/{id}": {
    "get": {
    "tags": [
    "Artist"
    ],
    "summary": "Get artist by ID",
    "operationId": "getArtistById",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int32"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "$ref": "#/components/schemas/ArtistDTO"
    }
    }
    }
    }
    }
    },
    "put": {
    "tags": [
    "Artist"
    ],
    "summary": "Update artist",
    "operationId": "updateArtist",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int32"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/ArtistDTO"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "$ref": "#/components/schemas/ArtistDTO"
    }
    }
    }
    }
    }
    },
    "delete": {
    "tags": [
    "Artist"
    ],
    "summary": "Delete artist",
    "operationId": "deleteArtist",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int32"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "OK"
    }
    }
    },
    "patch": {
    "tags": [
    "Artist"
    ],
    "operationId": "patchArtist",
    "parameters": [
    {
    "name": "id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "integer",
    "format": "int32"
    }
    }
    ],
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Artist"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "$ref": "#/components/schemas/ArtistDTO"
    }
    }
    }
    }
    }
    }
    },
...
    "/songs/filter": {
    "post": {
    "tags": [
    "Song"
    ],
    "summary": "Get artist by dynamic filter",
    "operationId": "filterBy",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FilterStruct"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "type": "object"
    }
    }
    }
    }
    }
    }
    },
...
    "/genres/filter": {
    "post": {
    "tags": [
    "Genre"
    ],
    "summary": "Get artist by dynamic filter",
    "operationId": "filterBy_1",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FilterStruct"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "type": "object"
    }
    }
    }
    }
    }
    }
    },
    "/artists": {
    "get": {
    "tags": [
    "Artist"
    ],
    "summary": "Get all artists",
    "operationId": "getAllArtists",
    "parameters": [
    {
    "name": "name",
    "in": "query",
    "required": false,
    "schema": {
    "type": "string"
    }
    },
    {
    "name": "country",
    "in": "query",
    "required": false,
    "schema": {
    "type": "string"
    }
    },
    {
    "name": "age",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "format": "int32"
    }
    },
    {
    "name": "dateOfBirth",
    "in": "query",
    "required": false,
    "schema": {
    "type": "string"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/ArtistDTO"
    }
    }
    }
    }
    }
    }
    },
    "post": {
    "tags": [
    "Artist"
    ],
    "summary": "Create a new artist",
    "operationId": "createArtist",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Artist"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "$ref": "#/components/schemas/ArtistDTO"
    }
    }
    }
    }
    }
    }
    },
    "/artists/filter": {
    "post": {
    "tags": [
    "Artist"
    ],
    "summary": "Get artist by dynamic filter",
    "operationId": "filterBy_2",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FilterStruct"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "type": "object"
    }
    }
    }
    }
    }
    }
    },
...
    "/albums/filter": {
    "post": {
    "tags": [
    "Album"
    ],
    "summary": "Get album by dynamic filter",
    "operationId": "filterBy_3",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FilterStruct"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "type": "object"
    }
    }
    }
    }
    }
    }
    },
...
    "/artists/name/{name}": {
    "get": {
    "tags": [
    "Artist"
    ],
    "summary": "Get artist by name",
    "operationId": "getArtistByName",
    "parameters": [
    {
    "name": "name",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "OK",
    "content": {
    "*/*": {
    "schema": {
    "$ref": "#/components/schemas/ArtistDTO"
    }
    }
    }
    }
    }
    }
    },
...
    },
    "components": {
    "schemas": {
    "Album": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int32"
    },
    "title": {
    "type": "string"
    },
    "year": {
    "type": "string"
    },
    "artist": {
    "$ref": "#/components/schemas/Artist"
    },
    "description": {
    "type": "string"
    },
    "numberOfSongs": {
    "type": "integer",
    "format": "int32"
    },
    "url": {
    "type": "string"
    },
    "songs": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Song"
    }
    }
    }
    },
    "Artist": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int32"
    },
    "name": {
    "type": "string"
    },
    "dateOfBirth": {
    "type": "string",
    "format": "date-time"
    },
    "country": {
    "type": "string"
    },
    "age": {
    "type": "integer",
    "format": "int32"
    },
    "singleSongList": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Song"
    }
    },
    "albums": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Album"
    }
    }
    }
    },
    "Genre": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int32"
    },
    "name": {
    "type": "string"
    },
    "description": {
    "type": "string"
    },
    "year": {
    "type": "integer",
    "format": "int32"
    },
    "songList": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Song"
    }
    }
    }
    },
    "Song": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int32"
    },
    "title": {
    "type": "string"
    },
    "time": {
    "type": "integer",
    "format": "int32"
    },
    "url": {
    "type": "string"
    },
    "artists": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Artist"
    }
    },
    "album": {
    "$ref": "#/components/schemas/Album"
    },
    "genreList": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/Genre"
    }
    }
    }
    },
    "SongDTO": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int32"
    },
    "title": {
    "type": "string"
    },
    "time": {
    "type": "integer",
    "format": "int32"
    },
    "url": {
    "type": "string"
    },
    "artists": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "type": "integer",
    "format": "int32"
    }
    },
    "album": {
    "type": "integer",
    "format": "int32"
    },
    "genreList": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "type": "integer",
    "format": "int32"
    }
    }
    }
    },
...
    "AlbumDTO": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int32"
    },
    "title": {
    "type": "string"
    },
    "year": {
    "type": "string"
    },
    "artist": {
    "type": "integer",
    "format": "int32"
    },
    "description": {
    "type": "string"
    },
    "numberOfSongs": {
    "type": "integer",
    "format": "int32"
    },
    "url": {
    "type": "string"
    }
    }
    },
    "ArtistDTO": {
    "type": "object",
    "properties": {
    "id": {
    "type": "integer",
    "format": "int32"
    },
    "name": {
    "type": "string"
    },
    "dateOfBirth": {
    "type": "string",
    "format": "date-time"
    },
    "country": {
    "type": "string"
    },
    "age": {
    "type": "integer",
    "format": "int32"
    },
    "singleSongList": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/SongDTO"
    }
    },
    "albumList": {
    "uniqueItems": true,
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/AlbumDTO"
    }
    }
    }
    },
    "FilterStruct": {
    "type": "object",
    "properties": {
    "listOrderCriteria": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/SortCriteria"
    }
    },
    "listSearchCriteria": {
    "type": "array",
    "items": {
    "$ref": "#/components/schemas/SearchCriteria"
    }
    },
    "page": {
    "$ref": "#/components/schemas/Page"
    }
    }
    },
    "Page": {
    "type": "object",
    "properties": {
    "pageIndex": {
    "type": "integer",
    "format": "int32"
    },
    "pageSize": {
    "type": "integer",
    "format": "int32"
    }
    }
    },
    "SearchCriteria": {
    "type": "object",
    "properties": {
    "key": {
    "type": "string"
    },
    "operation": {
    "type": "string",
    "enum": [
    "EQUALS",
    "CONTAINS",
    "GREATER_THAN",
    "LESS_THAN"
    ]
    },
    "value": {
    "type": "string"
    }
    }
    },
    "SortCriteria": {
    "type": "object",
    "properties": {
    "sortBy": {
    "type": "string"
    },
    "valuesorOrder": {
    "type": "string",
    "enum": [
    "ASC",
    "DESC"
    ]
    }
    }
    }
    }
    }
    }